Add GraphOfGrid partitioner to CpGrid::scatterGrid options and fix linking issues.
